requirements.txt
discord.py[voice]
ffmpeg
python-dotenv
